requirements.txt
asgiref==3.3.0
beautifulsoup4==4.9.3
Django==3.0.2
django-bootstrap4==2.3.0
django-tinymce==3.1.0
psycopg2==2.8.6
pytz==2020.1
soupsieve==2.0.1
sqlparse==0.4.1
